requirements.txt
Django==1.6.1
MySQL-python
pytz==2013.8
Pillow==2.8.2
PyJWT==1.4.0
elasticsearch==1.6.0
elasticsearch-dsl==0.0.8
